THC_Core: fix NaNs with FV and ideal EOS
Fix a bug that produced NaN when running prim_to_all to compute hydro
quantities at the cell interfaces near vacuum. The density was forced to
be larger or equal to the minimum value allowed by the EOS. For nuclear
EOS this value is positive, but for ideal EOS it is zero, resulting in
the enthalpy becoming NaN. The fix is to set the interface to atmosphere
if the density becomes too small (or negative).